[JBTM-2062] Move datasource bindings and txn mgr factory into the test package
