+++ ArjunaJTA/spi/src/test/java/io/narayana/spi/util/JndiProvider.java
@@ -21,10 +21,12 @@
  */
 package io.narayana.spi.util;
 
+import io.narayana.spi.InitializationException;
 import org.jnp.server.NamingBeanImpl;
 
 import javax.naming.Context;
 import java.util.Hashtable;
+import java.util.Map;
 
 
 public class JndiProvider {
@@ -47,5 +49,21 @@ public class JndiProvider {
 
     public static void stop() {
         jnpServer.stop();
+    }
+
+    public static void initBindings() throws InitializationException {
+        Map<String, DbProps> dbConfigs = new DbProps().getConfig(DbProps.DB_PROPERTIES_NAME);
+        DataSourceManagerImpl dataSourceManager = new DataSourceManagerImpl();
+
+        for (DbProps props : dbConfigs.values()) {
+            String url = props.getDatabaseURL();
+
+            if (url != null && url.length() > 0)
+                dataSourceManager.registerDataSource(props.getBinding(), props.getDriver(), url,
+                        props.getDatabaseUser(), props.getDatabasePassword());
+            else
+                dataSourceManager.registerDataSource(props.getBinding(), props.getDriver(), props.getDatabaseName(),
+                        props.getHost(), props.getPort(), props.getDatabaseUser(),props.getDatabasePassword());
+        }
     }
 }
