+++ ArjunaJTA/spi/src/test/java/io/narayana/spi/util/XADSWrapper.java
@@ -19,7 +19,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package io.narayana.spi.internal;
+package io.narayana.spi.util;
 
 import com.arjuna.ats.jdbc.TransactionalDriver;
